Separated the form and table into two different components inline with Single responsibility principle
